+++ examples/bs4demo/urls.py
@@ -4,10 +4,11 @@ from django.urls import include, path, re_path
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.views.generic import TemplateView
-from cmsplugin_cascade.cascade_theme.urls import extra_cascade_theme_patterns
+from cmsplugin_cascade.theme.urls import extra_cascade_theme_patterns
 
 class CascadeDemoView(TemplateView):
     template_name = 'bs4demo/strides.html'
+
 
 admin.autodiscover()
 
@@ -19,4 +20,3 @@ urlpatterns = [
     path('', include('cms.urls')),
 ]
 urlpatterns.extend(static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT))
-
