+++ examples/package.json
@@ -36,4 +36,3 @@
   },
   "homepage": "https://github.com/jrief/djangocms-cascade#readme"
 }
-
